+++ riakcs/CHANGELOG.md
@@ -2,7 +2,20 @@
 
 <!-- towncrier release notes start -->
 
-## 4.8.0 / 2025-09-05
+## 4.10.0 / 2025-11-26 / Agent 7.74.0
+
+***Added***:
+
+* Bump minimum version of datadog-checks-base to 37.24.0 ([#21945](https://github.com/DataDog/integrations-core/pull/21945))
+
+## 4.9.0 / 2025-10-02 / Agent 7.72.0
+
+***Added***:
+
+* Bump Python to 3.13 ([#21161](https://github.com/DataDog/integrations-core/pull/21161))
+* Bump datadog-checks-base to 37.21.0 ([#21477](https://github.com/DataDog/integrations-core/pull/21477))
+
+## 4.8.0 / 2025-09-05 / Agent 7.71.0
 
 ***Added***: