+++ slurm/CHANGELOG.md
@@ -2,6 +2,20 @@
 
 <!-- towncrier release notes start -->
 
+## 2.2.0 / 2025-11-26 / Agent 7.74.0
+
+***Added***:
+
+* Add process-level tags to scontrol metrics ([#21822](https://github.com/DataDog/integrations-core/pull/21822))
+* Bump minimum version of datadog-checks-base to 37.24.0 ([#21945](https://github.com/DataDog/integrations-core/pull/21945))
+
+## 2.1.0 / 2025-10-02 / Agent 7.72.0
+
+***Added***:
+
+* Bump Python to 3.13 ([#21161](https://github.com/DataDog/integrations-core/pull/21161))
+* Bump datadog-checks-base to 37.21.0 ([#21477](https://github.com/DataDog/integrations-core/pull/21477))
+
 ## 2.0.2 / 2025-07-10 / Agent 7.69.0
 
 ***Fixed***: